+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added `pageStateChange` event to `ViewModel`, triggered when `pageId`, `pageRef`, or `pageParams` changes
 - Added `query` property to functions produced from a string (jora query), i.e. with `Model#queryFn(query)`
 - Added `app-header` view
+- Added support for uploading files from `paste` event
 - Added `@discoveryjs/discovery/utils` export
 - Changed `nav-button` view to align with `button` view (`text`, `href` and `external` as config options, `disabled` is no longer treated as query)
 - Modified `PageRenderer#define()` and `ViewRenderer#define()` to allow specifying all options with a single config parameter:
@@ -16,7 +17,6 @@
 - Removed `export default` in the core modules
 - Fixed `Model#pathToQuery()` method to correctly generate a query when part of the path contains a jora keyword
 - Fixed an edge case in the `table` view when rendering a single cell with an explicit column setup and row data containing non-object values
-- 
 
 ## 1.0.0-beta.88 (24-10-2024)